🔥 remove: 알람로직 삭제
