+++ src/app/admin/volunteer/event/page.tsx
@@ -1,7 +1,7 @@
-import MyVolEventPage from '@/components/mypage/EventHistory/MyVolEventPage';
 import { COOKIE_ACCESS_TOKEN_KEY } from '@/constants/cookieKeys';
 import decodeDangleToken from '@/utils/token/decodeDangleToken';
 import { cookies } from 'next/headers';
+import MyVolEventPage from './MyVolEventPage';
 
 export default function VolunteerHistoryPage() {
   const accessToken = cookies().get(COOKIE_ACCESS_TOKEN_KEY)?.value || '';
