+++ src/components/mypage/MyPageMain/MyPageMain.css.ts
@@ -17,27 +17,6 @@ export const grid = style({
   marginLeft: '4px',
   cursor: 'pointer'
 });
-export const settingSection = style({
-  display: 'border-box',
-  width: `100%`,
-  maxWidth: BREAK_POINT,
-  marginLeft: `-${GLOBAL_PADDING_X}px`,
-  backgroundColor: palette.white,
-  margin: '8px 0 0 0'
-});
-
-export const txtGird = style({
-  display: 'flex',
-  flexDirection: 'column',
-  justifyContent: 'center',
-  alignItems: 'center'
-});
-export const stroke = style({
-  width: '1px',
-  height: '60px',
-  backgroundColor: palette.gray200,
-  margin: '0 22px 0 22px'
-});
 export const box = style({
   width: `calc(100% - 4 * ${GLOBAL_PADDING_X}px)`,
   marginTop: '20px',
@@ -50,6 +29,27 @@ export const box = style({
   borderRadius: '8px',
   cursor: 'pointer'
 });
+export const txtGird = style({
+  display: 'flex',
+  flexDirection: 'column',
+  justifyContent: 'center',
+  alignItems: 'center'
+});
+export const stroke = style({
+  width: '1px',
+  height: '60px',
+  backgroundColor: palette.gray200,
+  margin: '0 22px 0 22px'
+});
+
+export const settingSection = style({
+  display: 'border-box',
+  width: `100%`,
+  maxWidth: BREAK_POINT,
+  marginLeft: `-${GLOBAL_PADDING_X}px`,
+  backgroundColor: palette.white,
+  margin: '8px 0 0 0'
+});
 
 export const accountTxt = style({
   display: 'flex',
@@ -57,7 +57,6 @@ export const accountTxt = style({
   columnGap: '8px',
   alignItems: 'center'
 });
-
 export const accountBox = style({
   padding: '16px 20px 16px 20px',
   display: 'flex',
@@ -66,7 +65,6 @@ export const accountBox = style({
   alignItems: 'center',
   cursor: 'pointer'
 });
-
 export const noti = style({
   display: 'block',
   padding: '4px 20px 16px 20px'
@@ -76,4 +74,17 @@ export const divider = style({
   backgroundColor: palette.gray200,
   height: '1px',
   margin: '8 0 8 0'
+});
+export const loginContainer = style({
+  backgroundColor: palette.white
+});
+export const btnTxt = style({
+  textAlign: 'center'
+});
+export const btnWrapper = style({
+  display: 'flex',
+  flexDirection: 'column',
+  alignItems: 'center',
+  marginTop: '66px',
+  padding: '12px 6px 12px 6px'
 });
